Bug Fixes and Cart Integration
-- Test

+++ backend/src/main/java/com/glamourjewellery/glamour_jewellery/repo/CartRepo.java
@@ -3,7 +3,7 @@ package com.glamourjewellery.glamour_jewellery.repo;
 import com.glamourjewellery.glamour_jewellery.entity.Cart;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CartRepo extends JpaRepository<Cart, Integer> {
+public interface CartRepo extends JpaRepository<Cart, Long> {
+
 
 }
-
